chore: #455 set threshold as BigDecimal value table 'committee' and add tx_index to table committee_registration/deregistration

+++ aggregates/governance-aggr/src/main/java/com/bloxbean/cardano/yaci/store/governanceaggr/processor/ProposalStatusProcessor.java
@@ -489,7 +489,7 @@ public class ProposalStatusProcessor {
                     // get ratification result
                     RatificationResult ratificationResult = GovActionRatifier.getRatificationResult(
                             govActionDetail, isBootstrapPhase, maxAllowedVotingEpoch, ccYesVote, ccNoVote,
-                            BigDecimal.valueOf(ccThreshold), spoYesStake, spoAbstainStake, totalPoolStake,
+                            ccThreshold, spoYesStake, spoAbstainStake, totalPoolStake,
                             dRepYesStake, dRepNoStake, ccState, lastEnactedGovActionIdWithSamePurpose, isActionRatificationDelayed,
                             treasury,
                             EpochParam.builder()
