+++ stores/governance/src/main/java/com/bloxbean/cardano/yaci/store/governance/domain/Committee.java
@@ -8,6 +8,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 
 @Data
@@ -22,6 +23,6 @@ public class Committee {
     private Integer govActionIndex;
     private BigInteger thresholdNumerator;
     private BigInteger thresholdDenominator;
-    private Double threshold;
+    private BigDecimal threshold;
     private Long slot;
 }
